+++ src/rt/color.rs
@@ -5,13 +5,13 @@ use {
 
 #[derive(Debug, Clone)]
 pub struct Color {
-    red: f64,
-    green: f64,
-    blue: f64,
+    red: f32,
+    green: f32,
+    blue: f32,
 }
 
 impl Color {
-    pub const fn new(red: f64, green: f64, blue: f64) -> Self {
+    pub const fn new(red: f32, green: f32, blue: f32) -> Self {
         Self { red, green, blue }
     }
 
@@ -48,17 +48,17 @@ impl Color {
     }
 
     /// Returns the red component.
-    pub const fn red(&self) -> f64 {
+    pub const fn red(&self) -> f32 {
         self.red
     }
 
     /// Returns the green component.
-    pub const fn green(&self) -> f64 {
+    pub const fn green(&self) -> f32 {
         self.green
     }
 
     /// Returns the blue component.
-    pub const fn blue(&self) -> f64 {
+    pub const fn blue(&self) -> f32 {
         self.blue
     }
 }
@@ -95,10 +95,10 @@ impl Sub for Color {
     }
 }
 
-impl Mul<f64> for Color {
+impl Mul<f32> for Color {
     type Output = Self;
 
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Self {
             red: self.red * rhs,
             green: self.green * rhs,
@@ -107,10 +107,10 @@ impl Mul<f64> for Color {
     }
 }
 
-impl Mul<f64> for &Color {
+impl Mul<f32> for &Color {
     type Output = Color;
 
-    fn mul(self, rhs: f64) -> Self::Output {
+    fn mul(self, rhs: f32) -> Self::Output {
         Color {
             red: self.red * rhs,
             green: self.green * rhs,
